Refactor auth modules and add /auth prefix
