+++ public/auth/login/index.html
@@ -6,7 +6,7 @@
     <title>Document</title>
   </head>
   <body>
-    <form action="/login" method="post">
+    <form action="/auth/login" method="post">
       <div>
         <label>Username:</label>
         <input type="text" name="username" />
