+++ routes/index.js
@@ -1,11 +1,9 @@
 var express = require("express");
-const passport = require("passport");
-const authController = require("../controllers/authController");
 var router = express.Router();
+const authController = require("../controllers/authController");
 
 router.use("/api", require("./api"));
-router.post("/login", passport.authenticate("local"), authController.loginPost);
-router.post("/logout", authController.logoutPost);
+router.use("/auth", require("./auth"));
 router.post("/user", authController.createUser); // TODO - move somewhere else
 
 module.exports = router;
